[REFACTOR] 폴더링 구조 수정
