+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.home.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="false" />
     </application>
