+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/FriendProfile.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.presentation.home
+package org.sopt.dosopttemplate.presentation.main.home
 
 import androidx.annotation.DrawableRes
