+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.presentation.home
+package org.sopt.dosopttemplate.presentation.main.home
 
 import androidx.lifecycle.ViewModel
 import org.sopt.dosopttemplate.R
