+++ app/src/main/res/layout/fragment_android.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.home.AndroidFragment">
+        tools:context=".presentation.main.AndroidFragment">
 
         <!-- TODO: Update blank fragment layout -->
         <TextView
